Basic source spans
Loc is now (usize, usize), representing a pair of indices into the
source code. We use this to render a wiggly line under the locus of
errors. Currently we assume that the line doesn't cross newlines.

We also have slight problems where the span can run over into trailing
whitespace.